Move tunnel log rotation into tunnel monitoring module
